Refactor wordlist
